closer still
Now only generator unit tests and template applicator unit tests.
