+++ DnDGen.CreatureGen.Tests.Unit/Templates/LycanthropeApplicatorApplyToAsyncTests.cs
@@ -49,7 +49,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
                 .Returns(baseCreature.Demographics);
         }
 
-        [TestCaseSource(nameof(BUG_HitPointTotals))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.BUG_HitPointTotals))]
         public async Task BUG_ApplyToAsync_AddAnimalHitPoints_NegativeConstitutionBonus(int bQ, int bD, int bR, double bA, int aQ, int aD)
         {
             baseCreature.HitPoints.HitDice[0].Quantity = bQ;
@@ -94,7 +94,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
                 $"Base roll: {creature.HitPoints.HitDice[0].DefaultRoll}; Base Roll: {baseRoll}; Animal Roll: {animalHitPoints.HitDice[0].DefaultRoll}, Bonus: {bonus}");
         }
 
-        [TestCaseSource(nameof(BUG_HitPointTotals))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.BUG_HitPointTotals))]
         public async Task BUG_ApplyToAsync_AddAnimalHitPoints_WithConstitutionBonus(int bQ, int bD, int bR, double bA, int aQ, int aD)
         {
             baseCreature.HitPoints.HitDice[0].Quantity = bQ;
@@ -154,7 +154,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
                 Throws.InstanceOf<InvalidCreatureException>().With.Message.EqualTo(message.ToString()));
         }
 
-        [TestCaseSource(nameof(IncompatibleFilters))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.IncompatibleFilters))]
         public async Task ApplyToAsync_ThrowsException_WhenCreatureNotCompatible_WithFilters(
             bool asCharacter,
             string type,
@@ -699,7 +699,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(animalAttacks[1].Name, Is.EqualTo("animal attack 2 (in Animal form)"));
         }
 
-        [TestCaseSource(nameof(SizeComparisons))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.SizeComparisons))]
         public async Task ApplyToAsync_AddLycanthropeAttacks_BaseIsBigger(string smallerSize, string biggerSize)
         {
             baseCreature.Size = biggerSize;
@@ -731,7 +731,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(lycanthropeAttacks[1].Name, Is.EqualTo("lycanthrope attack 2"));
         }
 
-        [TestCaseSource(nameof(SizeComparisons))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.SizeComparisons))]
         public async Task ApplyToAsync_AddLycanthropeAttacks_AnimalIsBigger(string smallerSize, string biggerSize)
         {
             baseCreature.Size = smallerSize;
@@ -763,7 +763,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(lycanthropeAttacks[1].Name, Is.EqualTo("lycanthrope attack 2"));
         }
 
-        [TestCaseSource(nameof(Sizes))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.Sizes))]
         public async Task ApplyToAsync_AddLycanthropeAttacks_AnimalAndBaseAreSameSize(string size)
         {
             baseCreature.Size = size;
@@ -796,7 +796,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(lycanthropeAttacks[1].Name, Is.EqualTo("lycanthrope attack 2"));
         }
 
-        [TestCaseSource(nameof(Sizes))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.Sizes))]
         public async Task ApplyToAsync_AddLycanthropeAttacks_WithBonuses(string size)
         {
             baseCreature.Size = size;
@@ -1425,7 +1425,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(creature.Saves[SaveConstants.Will].IsConditional, Is.True);
         }
 
-        [TestCaseSource(nameof(ChallengeRatings))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.ChallengeRatings))]
         public async Task ApplyToAsync_IncreaseChallengeRating(string originalChallengeRating, double animalHitDiceQuantity, string updatedChallengeRating)
         {
             baseCreature.ChallengeRating = originalChallengeRating;
@@ -1437,7 +1437,7 @@ namespace DnDGen.CreatureGen.Tests.Unit.Templates
             Assert.That(creature.ChallengeRating, Is.EqualTo(updatedChallengeRating));
         }
 
-        [TestCaseSource(nameof(LevelAdjustments))]
+        [TestCaseSource(typeof(LycanthropeTestData), nameof(LycanthropeTestData.LevelAdjustments))]
         public async Task ApplyToAsync_IncreaseLevelAdjustment(int? oldLevelAdjustment, int? newLevelAdjustment, bool isNatural)
         {
             baseCreature.LevelAdjustment = oldLevelAdjustment;
